making changes to include profile options.

+++ benchmarks/blackscholes/blackscholes.c
@@ -109,7 +109,7 @@ fptype CNDF ( fptype InputX )
     xInput = InputX;
 
     // Compute NPrimeX term common to both four & six decimal accuracy calcs
-    expValues = exp(-0.5f * InputX * InputX);
+    expValues = exp((double)(-0.5f * InputX * InputX));
     xNPrimeofX = expValues;
     xNPrimeofX = xNPrimeofX * inv_sqrt_2xPI;
 
@@ -181,9 +181,9 @@ fptype BlkSchlsEqEuroNoDiv( fptype sptprice,
     xVolatility = volatility;
 
     xTime = time;
-    xSqrtTime = sqrt(xTime);
+    xSqrtTime = sqrt((double)xTime);
 
-    logValues = log( sptprice / strike );
+    logValues = log( (double)(sptprice / strike) );
 
     xLogTerm = logValues;
 
@@ -205,7 +205,7 @@ fptype BlkSchlsEqEuroNoDiv( fptype sptprice,
     NofXd1 = CNDF( d1 );
     NofXd2 = CNDF( d2 );
 
-    FutureValueX = strike * ( exp( -(rate)*(time) ) );
+    FutureValueX = strike * ( exp( -(double)(rate)*(time) ) );
     if (otype == 0) {
         OptionPrice = (sptprice * NofXd1) - (FutureValueX * NofXd2);
     } else {
